CX-411: Cleaned up code

+++ source/plugins/org.custusx.registration.method.landmark/cxLandmarkRegistrationWidget.cpp
@@ -208,18 +208,18 @@ void LandmarkRegistrationWidget::activateLandmark(QString uid)
   */
 QString LandmarkRegistrationWidget::getNextLandmark()
 {
-		std::vector<Landmark> lm = this->getAllLandmarks();
+	std::vector<Landmark> lm = this->getAllLandmarks();
 
-		int size = int(lm.size());
-		for (unsigned i=0; int(i) < size-1; ++i)
-    {
-        if (lm[i].getUid()==mActiveLandmark)
-        {
-            return lm[i+1].getUid();
-        }
-    }
+	unsigned long size = lm.size();
+	for (unsigned i=0; i < size-1; ++i)
+	{
+		if (lm[i].getUid()==mActiveLandmark)
+		{
+			return lm[i+1].getUid();
+		}
+	}
 
-    return "";
+	return "";
 }
 
 std::vector<Landmark> LandmarkRegistrationWidget::getAllLandmarks() const
